refactor to devops service
